Autoload ocamlfind from "command-env".
Summary: The [`esy`](https://github.com/esy/esy) project manager outputs
a `command-env` that describes all the `ocamlfind` dependencies, and
dependencies.

What this diff accomplishes:
- Makes VSCode or any LSP based Reason/OCaml project automatically work
without any configuration.
- Just open VSCode on any `esy` project and autocompletion will just
work.

How: All this diff does is wrap all the calls like this
- `command-env-exec ocamlmerlin`
- `command-env-exec refmt --print-width 80`
etc.

Note: We haven't yet began to emit `command-env-exec` from `esy`, but we
will very soon. We wanted to get this out there so that it will be ready
for people.

Note: The `command-env-exec` wrapper is not `esy` specific and any
package/project manager out there could emit this file.

I tested on a project with a `command-env-exec`, and another project
without, just to make sure everything still works.