APIs For Introspecting Reason/OCaml LSP Environments
**This diff implement Two New Language Server Features For Exploring IDE Environment.**

One common issue with the vscode Reason plugin is that people are running it in an environment that consists of some hybrid of Opam, npm etc without fully understanding their dev time dependencies or what LSP "sees". This makes it difficult to help the Discord community.

I've added a server API to explore the user's environment, with an autocomplete window to search for various keywords such as `"merlin_version"`, or `PATH`.  When hitting enter on an entry, it will open the variable in a new editing buffer. The user may also choose `Show Entire Environment` which will open up the environment that the language server sees in a new windows, so that users may copy/paste it into the Discord channel so that we can help debug their setup.

Another issue is that when building native libraries, such as the Reason parser itself, the build is misconfigured or tries to depend on libraries that aren't actually provided by the package dependencies.  Each ocaml package contains multiple possible library names - it's not a 1-1 mapping, so it can be confusing to people new to the ecosystem.  Here, I'm also adding a feature to display a popup list of ocaml *library* dependencies provided by the project dependencies. Because of [my previous PR](https://github.com/freebroccolo/ocaml-language-server/pull/44) all of these introspections are now project specific. It will fall back to global introspection if you don't have well-isolated projects.